chain rstudio panel: use Accordion component for it
