+++ webapp/views/App/views/Analysis/Chain/ChainDetails.scss
@@ -28,7 +28,7 @@
       }
     }
 
-    .rstudio-fieldset {
+    .rstudio-accordion {
       margin: auto;
 
       .content {
